Correctly tag provenance in array_like_{newelem,set}
Summary:
This was two old bugs in both array_like_newelem and array_like_set that
would cause us to inappropriately propagate the provenance tag of the input
array, mostly from missed cases when empty+tagged arrays became a possibility
in these fns.

If we are adding an element to an array-like union that is possibly empty or
nonempty, with a known provenance tag, it's incorrect to propagate the
provenance of the arrlikedata, since it corresponds _only_ to the nonempty part
of the union. (And the empty part of the union could have any provenance tag.)
So in this scenario, we need to discard the provenance tag.

We would also unconditionally use the source location if the input array was
empty in `newelem`, which was wrong after we started allowing empty arrays w/
provenance tags

Reviewed By: mxw

Differential Revision: D18530971

fbshipit-source-id: bdc35ed028c489360fec0b8d9205133f072aced4