Remove ignore_pos option
Summary:
Compiling files using the [`concurrent`](https://docs.hhvm.com/hack/asynchronous-operations/concurrent) feature fails with cryptic assertion failure.
```
.main {
  String "\"Assert_failure .../full_fidelity_ast.ml:251:8\""
  Fatal Parse
}

```
which refers to the following:
```
  let process_lifted_awaits lifted_awaits =
    List.iter lifted_awaits.awaits ~f:(fun (_, (pos, _)) ->
        assert (pos <> Pos.none));
    List.sort
      lifted_awaits.awaits
      ~compare:(fun (_, (pos1, _)) (_, (pos2, _)) -> Pos.compare pos1 pos2)

```
User needs to remember to enable the `hack.compiler.sourcemapping` or `eval.disassemblersourcemapping` option in order to compile `concurrent` correctly. This has the result of also printing the source_loc lines.
## This diff attempts to enable concurrent by default while also making it unnecessary to include the source locs in bytecode by deleting the option to ignore pos while keeping `hack.compiler.sourcemapping` and `eval.disassemblersourcemapping` controlling the emission of source locs in bytecode.
There are several other places in the codebase where we needed positional information (XHP and `__LINE__` const). We would remember positions for those times. We might be able to also do a similar thing for concurrent - but I feel that'd only be necessary if performance is really taking a hit, which I can't imagine it is since we are doing this anyway for `concurrent` in production.

Reviewed By: shiqicao

Differential Revision: D18431214

fbshipit-source-id: 41ca987ea51244b22e35cdd0691fe30914e995ef