Move optimization for function parameters
Summary:
We now throw an exception when a caller passes too few arguments to a callee. See: throw_missing_arguments_nr, which can be caught by the caller but not by the callee. As a result, HHBBC can assume that parameters are initialized.

Guaranteeing that a local is initialized allows for move optimizations, like using PushL rather than CGetL to avoid refcount operations for their last use.

This change broke a number of tests, because we'd do PushL optimizations of parameters that could be observed when EnableArgsInBacktraces is set. The correct behavior is to mark all locals live when we encounter a bytecode that may throw and this flag is set. We make this change in this diff rather than in the previous one in the stack because it's an actual logical change, and we don't want to mix it in with test-config changes.

Reviewed By: markw65

Differential Revision: D18603704

fbshipit-source-id: 5adbad4bd1941fa507379f0cc67b2b07501eafd2