presize NamedEntityPairTable in repo mode
Summary:
Add a `GetLitstrCountStmt`, which is used before loading all string literals from the repo. It avoids reallocation in the compact vector, and prepares for lazy loading.

Also the *Stmt's don't have to be member of LitstrRepoProxy, as they are mostly used only once per process.  Create them on the stack when needed.

Reviewed By: markw65

Differential Revision: D18662324

fbshipit-source-id: 96b3458f0c4454fcbee218751f515244fe365df8