Fix IDE speed on IViewerContext
Summary:
As reported by Dave Alongi, some users reported significant slowdowns when browsing through certain complex root-level files.

The root cause turned out to be that a spurious file change notification.  The file change notification was for a file that was not a hack file; but our code wiped out the memoized file context and caused the file to be rescanned.

Also: Previously, we were reading the file off disk to send its contents to IDE services.  There's no need to do this, since the IDE services will read the file off disk when it needs them.

Reviewed By: arxanas

Differential Revision: D18667389

fbshipit-source-id: 1bef15ce0bea0f51d6eea60b5cad3e20e316f23b