Ignore decrefs in load-elim
Summary:
DecRefs no longer re-enter, so the only thing a DecRef can do is free
memory; from load-elim's point of view this is equivalent to a
GeneralEffects with a kill set, but no loads or stores. We can't model
it that way in the memory effects, because some passes really do care
about the DecRef's effects (store-elim needs to make sure that it
doesn't kill any PureStores to an array thats about to be freed, or we
could double free the old values, for example).

So lets just special case DecRef and DecRefNZ in load-elim, and ignore
them.

Reviewed By: ricklavoie

Differential Revision: D18658474

fbshipit-source-id: f44b857df981be96d3faae7145496639b200df7c