Add setMove for collections
Summary:
We don't need to inc-ref the value in MapSet and VectorSet. Add methods that let us skip these RC ops.

We can clean this code up further, but the best next step is to first get rid of the Variant setters (which obscure refcounting) for both array-likes and collections.

Reviewed By: mofarrell

Differential Revision: D18642955

fbshipit-source-id: c4666a3a05cb900e37ea43dc3b7e8dbbc020c2bf