serverFormat cleanup
Summary:
I'm writing this diff because (1) serverFormat uses some Lsp types, (2) I tried to change it to use canonical Lsp methods to manipulate those types, (3) doing this caused a linker error because it brought in Lwt into server.

I'm not sure how exactly it brought in Lwt. But it was enough for me to look, and see that Lsp data (particularly editor_open_files) have no role in serverFormat and should be removed.

Prior to this diff, the function `go_ide` took in all editor_open_files. It took them in so it could extract the file contents of the file we're trying to operate upon. It knew which file we're operating upon because the "action" included the filename (or was it the URI? unclear).

All of this is a pretty convoluted way to achieve what we actually want, which is merely to deliver the file-content!

I figure that clientLsp is the only party that should be aware of editor_open_files, and aware of URIs.

Hence: this diff changes serverFormat.go_ide so it takes in the content directly, and relies upon clientLsp to gather the content first.

I also renamed "filename" to "filename_for_logging" to make clear that it has no actual material benefit.

Reviewed By: arxanas

Differential Revision: D18654256

fbshipit-source-id: 9134f061965a1bd5372efc2c518adb8f50c4dcdb