Add support for indirect fixups in unwinder
Summary:
Do not construct fake ActRec. Instead of obtaining the FP register of the
caller and building a fake ActRec, obtain the CFA pointer, which points to the
high end of the native callee frame. Reduce it by kNativeFrameSize to obtain
the real ActRec, which works with indirect fixups.

It should be possible to simplify this further and eliminate _Unwind_GetIP()
calls from the rest of the unwinder, but I wanted to first assert that what we
get from the CFA is the same.

Reviewed By: oulgen

Differential Revision: D18669875

fbshipit-source-id: 623a4c891df3ddeb07de2c6060e4508a8d7d0924