Initial commit of GDB indexer
Summary:
This diff adds a custom GDB indexer using our own DWARF parser. GDB takes 3.5 minutes to index HHVM binary. The goal of this diff is to improve that speed. Built using `mode/opt-hhvm-lto`, this indexer is able to generate the index file in 14 seconds. Through more optimizations, we will be able to further bring this number down.

Source for the implementation:
https://sourceware.org/gdb/current/onlinedocs/gdb/Index-Section-Format.html

Reviewed By: jano

Differential Revision: D18576535

fbshipit-source-id: 4cf72a1f93d2c722c1ea65f9d60caf61a227a878