Add fixups to Calls
Summary:
Prologues will need an ability to throw exceptions prior to spilling ActRec,
when they still operate in the stublogue mode. This means either we need to
eagerly sync registers before C++ calls that may throw, which would cause
significant code size increase and is non-trivial, as the caller may be an
inlined frame, or we can rely on indirect fixups from the stublogue mode of
prologues to point to fixups registered with Calls.

So, let's do that and register fixups with Calls.

This change was originally a part of D18095870.

Reviewed By: oulgen

Differential Revision: D18691963

fbshipit-source-id: d8b018e318e0e9afab323de4b6f9c19302c14797