serverFindRefs logging
Summary:
I'm trying to track down a heisenbug that I observe on test machines running some --find-refs tests, but not on others.

This diff solely adds a few Hh_logger.debug statemnets in strategic places inside the implementation of --find-refs. In normal course of affairs these don't do anything. But if the test does `hh --verbose-on` beforee `hh --find-refs`, then all of these debug statements will now appear in the server log. That'll be what I need to debug.

Reviewed By: arxanas

Differential Revision: D18718955

fbshipit-source-id: 84097cd88667c92ec6a9ad54901d7ac94cda3c29