Add APIs for copying and reading slabs
Summary:
In order to implement [zero-compression, zero-copy sharing of decls](https://fb.quip.com/Z09yA6WOQguA#fcCACAitztE), we need the ability to write OCaml values into contiguous regions of memory we do not manage. The `slab` module was added for this purpose, but still only provides OwnedSlabs (which own the region of memory containing the value, and are useful for tests and mocks). We need a means of writing slabs into slices and reading slabs from slices. Not only that, since cachelib does not provide any guarantees about alignment, we need to write and read slabs embedded in *unaligned byte slices*.

This diff does so by allocating 7 extra bytes whenever we reserve memory for a slab. We treat the word-aligned contents of the byte string as the slab, and ignore the leading and trailing padding bytes.

Reviewed By: jewelpit

Differential Revision: D18671041

fbshipit-source-id: 8d1829f3b7cde5aa7fb2392223c8277753349637