Hg.current_mergebase_hg_rev : use a reliable query
Summary:
The query I'm replacing returns nothing if you're already on a public commit. The replacement query always returns the correct result. According to the comments in `get_closest_global_ancestor`, it's true even when we're in the middle of a rebase with merge conflicts.

I'm also removing the `~plain` parameter from `exec_hg` because it was only used in the old query code that I replace here.

Reviewed By: Wilfred

Differential Revision: D18713788

fbshipit-source-id: 060b1edb89f905ba406e4f6b3368aefd2f19b92c