index provider doesn't need to iterate all files
Summary:
It takes us 30seconds during lazy init to iterate all files. The only reason we do that is so the symbol-indexer will later (asynchronously) be able to work with them.

That's kind of weird, that we do a costly sync in order to allow a later asynchronous thing!

It's also weird because the behavior of the asynchronous work (in case of CustomIndex or NoIndex) is simply a no-op!

This iterating used to be skipped. It was re-added in D17434062 for the reason that "tests were failing". I'm going to skip it again, and see just which tests were failing.

Reviewed By: 2BitSalute

Differential Revision: D18721500

fbshipit-source-id: 29c28ec957917839e00865967bf72841a717ceb9