Fix extraction of type constants referenced through nested type accesses
Summary: When adding a dependency on a type constant of the form `C0::T1::T2::...TN`, we look up `T1` in `C0`, and if it is defined and points to a type `C1`, we recursively add a dependency on `C1::T2::...TN`.  We should also do this if `T1` is abstract in `C0` but is constraint by an upper bound `C1`.  (See the new test case below.)

Reviewed By: kmeht

Differential Revision: D18729826

fbshipit-source-id: 57769d479b5fd92be185065a5c0beaa8f1168621