Add --merge-global-tvenvs to hh_single_type_check
Summary:
Add a mode to hh_single_type_check to run global inference on a single file and merge the resulting global type variable environments. Print the merged environment, as well as merge errors, and type checking errors.

I don't particularly like this code because not all of its output might be relevant (lots of output for .hhi files), but I'n not keen on suppressing any of it, because it actually might be relevant.

Open question: **should we actually be generating global type variables for untyped .hhi files?** My intuition would be NO! And in that case, the problem of non-relevant output would be solved. I'm not sure whether we generate global type variables for .hhi files when running global inference on www. If not, we shouldn't be doing it here either, if so, we might need to reconsider that. *To be investigated...*

In the mean time this diff is still useful for debugging and deving.

Reviewed By: Wilfred

Differential Revision: D18729306

fbshipit-source-id: 697663a030adaca8775e09d3b8710bfe830986a0