Elaborate in a given namespace the names of classes / interfaces in `require extends` / `require implements` clauses
Summary:
I'm not sure why we don't do this already.  As a result,
```
class C<T> extends T {}
```
is a naming errors (unbound name) while
```
interface I<T> {
  require extends T;
}
```
is a type error.

(A related issue: because of elaboration of the name `T` in the former example to `\T`, the check if this name occurs in the list of type parameters of the class fails because the type parameter names are not elaborated (duh), but in the latter example because `T` is not elaborated the check succeeds and `T` is transformed from `Happly ((_, "T"), [])` into `Habstr "T"`.)

Reviewed By: kmeht

Differential Revision: D18749656

fbshipit-source-id: 1b56c52840d2ff331e10ebb5a7dce895c82727ce