Relocate in parallel with translate
Summary:
The relocation step of retranslateAll runs in a single thread, and is hard to parallelize because it has to lay the code out in order, but it could in theory start running as soon as we finish the first function.

This lets the main thread continue running, waiting for each function to be optimized before proceding to relocate it.

This looked like it should be easy, because the translate threads all translate into local buffers, and don't (seem to) require the code or metadata locks. But In jumpstart mode, the srcDB is empty to begin with, so we have to generate a new anchor translation for each region, and *that* requires the code and metadata locks. So we have to make the relocate thread minimize the amount of time it holds the locks to prevent blocking all the translate threads. The previous two diffs add the mechanism for doing so with prologue and funcbody translations, and this makes use of it, and does the same for relocateFunction too.

Differential Revision: D18647410

fbshipit-source-id: 4b29fdd0b5ba8a47131e6ab1a029d63fc7741716