Restructure prologue emission
Summary:
Currently, emitting a prologue requires holding the code and metadata locks for the whole operation, but they're only required for part of it.

Restructure the code a bit, and allow callers to pass in a CodeMetaLock to just lock for the critical section. A later diff in the stack will add uses of the CodeMetaLock parameter.

Reviewed By: binliu19

Differential Revision: D18535669

fbshipit-source-id: d9e5aebd21a6aba673999a16739f5c48a6cf019c