Include refcount opts in retry loop
Summary:
I noticed that after inlining, load/store elim can get blocked by
DecRefs - eg when we inline a constructor which sets some properties,
there's an IncRef/DecRef pair wrapping the $this of the inlined
constructor. The DecRef is marked as storing to AHeapAny, so
subsequent loads from the just-inited properties will be unknown;
however, if there *are* any such loads, the DecRef clearly didn't free
the Object, so at the least could have been a DecRefNZ with *no* heap
effects (and in most cases will be eliminated along with the matching
IncRef by recount opts).

Reviewed By: ottoni

Differential Revision: D18085629

fbshipit-source-id: b244d94cdf6e8c10ba249d97083b7c60bccf1e17