Use autocargo in hphp/hack
Summary:
Generate `Cargo.toml` files from `TARGETS` using [autocargo](https://fb.workplace.com/groups/rust.language/permalink/3035731399808788/). Important part is that to make it work I had to (as in "that was the easiest way to make it work") blow up our `Cargo.lock` and regenerate it completely from the scratch, which likely changed some versions of used libraries. I don't think there was any deep thought besides "it builds" when choosing previous versions, so I'm going to assume that as long as it still builds it's still fine.

There are still few manual Cargo files due to limitations of `autocargo`, but it get's us most of the way there.

Reviewed By: shiqicao

Differential Revision: D18769893

fbshipit-source-id: 2f48ab01166ffca895ff07d4f50315619d7dbc57