test for Decl_not_found
Summary:
We found a case where `hh --identify-function ... < contents` was generating Decl_not_found exception and failing to give any results.

It was a weird case because it involved two files, and each of them worked if you did them individually, but the second one failed if you did both of them.

This calls for a test case!

I've done it with a shell-script because our other test harnesses don't cover this case, including running two separate commands.

Reviewed By: arxanas

Differential Revision: D18738241

fbshipit-source-id: 695b0db725a58ec0c306a1b2eac34e8732b6132b