AST provider always uses ctx
Summary:
Prior to this diff, the "decl_from_stdin" was failing. Here's how:

(1) user did `hh --identify-symbol line:col < stdin1`. Because it got file contents from stdin, it didn't have a filename associated with it.

(2) ServerRpc handles FIND_REFS requests in the standard way by (1) creating a temporary contexts where the empty filename maps to stdin1 contents, (2) quarantining all the shared-mem heaps by using a local stack.

(3) But the act of creating a temporary context had the side effect of writing the ast from stdin1 into the ParserHeap. (Why did it do this? so that Ast_provider.get_ast was able to read it, I guess.) And because we weren't yet quarantined, we were writing it into the shared ParserHeap.

Invariant violation: the shared ParserHeap should never contain ephemeral ASTss.

(4) By the time the second `hh --identify-symbol line:col < stdin2` is invoked, we do the same. But this time our attempt to update the ParserHeap is a no-op since the old AST is already there under the same empty-filename key.

The test "test/integration:decl_from_stdin" embodied this scenario, and would result in a Decl_not_found error in the server log when the second time around attempts to get the AST for stdin2 and get the decl of a symbol from it. (it fails because it ends up getting the AST for stdin1).

**What this diff does:**

(1) Creating a temporary context no longer has the side effect of writing the AST into ParserHeap. Indeed there's no reason why it should. We already store the AST in the ctx, and that's the right canonical place for it to be.

(2) Concommitantly, `Ast_provider.get_ast` has been rewritten so it always looks first in ctx for an AST, no matter the provider. This indeed embodies the entire purpose of a ctx. I've added a few more comments too into the function.

**Additional context:**

A full 2.1% of hover and go-to-def requests are currently failing due to Decl_not_found: https://fburl.com/scuba/3eixhegw

* hh_server ready, fast response - 93.2%
* hh_server ready, slow response - 2.7%
* hh_server global typechecking, fast response - 1.1%
* hh_server global typechecking, slow response - 0.2%
* hh_server local typechecking, fast response - 0.4%
* hh_server local typechecking, slow response - 0.1%
* error response - 2.1% (all due to Decl_defs.decl_not_found, https://fburl.com/scuba/vldmlpra)

I have high hopes that this diff will solve them...

Reviewed By: arxanas

Differential Revision: D18740104

fbshipit-source-id: cdadae36dd491e5a28f3a9aeadedfe2c7c508dbf