switch Utils.dn to Hh_logger.debug in decl
Summary:
Utils.dn will print to stderr if the "Utils.debug" flag is set. That never gets set in the normal course of affairs; it's only ever set by a developer manually and then rebuild.

I've switched some uses of it over to Hh_logger.debug. The differences: (1) it goes to server log rather than stderr, so we can see it in context; (2) it's controllable by hh.conf and also by `hh --verbose-on` to turn it on. This seems better all around.

Reviewed By: 2BitSalute

Differential Revision: D18747153

fbshipit-source-id: dab3479755b2867e2910558a8888c7dc7fed6b10