remove log of parse time
Summary:
If the user passes `--profile-log`, this diff causes it to print to stderr messages of the form "<duration> <file> [parsing]"

This was only ever used for ad-hoc debugging. We don't gather or analyze the data.

So I'm going to remove this. I think we benefit from a proper form of telemetry around parsing time - looking at overall parse time, decl-parsing, other forms of parsing, focusing on the Rust time. But the bit of code I'm deleting wasn't that.

Reviewed By: CatherineGasnier

Differential Revision: D18747172

fbshipit-source-id: 8caa6bc9e06349930c4976dd3c476b42a13f6d4c