don't emit debugging info to stdout
Summary:
From the dawn of our codebase (2014) there's been a flag Utils.debug which a developer can set manually and then rebuild. One effect is that errors get shown on stdout by server, in addition to being displayed normally.

This is kind of goofy. No one ever uses it. I'm removing it.

Reviewed By: 2BitSalute

Differential Revision: D18747154

fbshipit-source-id: bfe475b427c349c506d1857175464993e5354c09