add profile_type_check_twice
Summary:
The current behavior of `--profile-log` is to emit a log of time to "fetch-necessary-decls-and-typecheck".

I want to leave the existing behavior, but also add a new option: `--config profile_type_check_twice=true`. This new option will cause it to (1) fetch necessary decls and typecheck, (2) typecheck a second time now that the necessary decls have been fetched. It will be slower, of course, but it will give us a pure typecheck time.

I'm adding this as an option because it's slower and therefore won't be able to be used on every single sandcastle run (as is currently the case).

This diff doesn't yet implement the behavior of `--config profile_type_check_twice=true`. This diff is merely to add the flag and plumb it through.

Reviewed By: 2BitSalute

Differential Revision: D18747163

fbshipit-source-id: 5a8b6f422a2be31707c6ad65790965bfc1848b2f