implement profile_type_check_twice
Summary:
This diff implements the flag `--config profile_type_check_twice=true`.

We plumb it through into check_info, which is consumed by Typing_check_service.process_files.

Its behavior is, after typing is done, to run typing a second time. The idea is that for this second time all the decls will already be fetched and in cache, so we're able to log a "pure" typing time.

I decided to log the raw time durations in telemetry (first duration is decl+typechecking, second duration is typechecking), rather than attempting subtract them to figure out how long was spent in decls and how long in typing. I figure that whoever's analyzing the data can perform that subtraction.

I added a new column "bytes_per_s". That's because the existing "filesize" and "time_taken_ms" are integers and have lost granularity - e.g. many `time_taken_ms` are rounded to 0ms or 1ms. I looked at outlier values and saw that "bytes_per_s" varies from ~500 to ~2M. So it's the right order of magnitude.

I decided that `--config profile_type_check_twice` will turn on accurate+costly "double typechecking", even in the absence of `--profile-log`. That's partly because I like having the command-line switches completely orthogonal, and partly because I want to measure the overhead of double-typechecking even in the absence of telemetry.

Reviewed By: 2BitSalute

Differential Revision: D18747186

fbshipit-source-id: f606af2bfd4f142e7108027fcee81feb5584bfaa