Catch errors when writing out symbol info json
Summary:
Previously, when I tried to run the typing symbol info writer to produce JSON
for all of www, it would die (often completely silently) when ~780 JSON files
had been fully/partially produced.

With some added logging in hh_server, I managed to get an error that indicated
the problem was an exception when computing the symbols, killing a worker
and in turn causing all the other workers/main thread to die (P123321507).

With the changes in this diff to catch these kinds of exceptions,
I was able to run the symbol indexer over www to completion for the first time.
The exception is printed to the output of the run.

surprise-dolphin

Reviewed By: jewelpit

Differential Revision: D18785711

fbshipit-source-id: 7b44b20c88920d04e4e9691e2ecd3eaa02dfc553