Simplify TypeStructure::resolve internals.
Summary:
resolveTS took a const Class::Const&, but none of the entry points to it had one - most were totally faking one, and even the code in class.cpp had to make a copy to avoid race conditions. Internally, resolveTS and its helpers were passing the const Class::Const& around for just 2 of the things in it: the name and the declaring class; they were also passing around the typeCnsCls and an array of generics, leading to a large list of arguments (which mostly just passed through to other helpers).

Create a TSCtx that can package up the 4 items being passed around in a single arg and without trying to reuse Class::Const in a silly way. This does not become part of the TSEnv because recursing through T_UNRESOLVED during resolution completely replaces the things in TSCtx but needs to continue to reference the same values in the TSEnv.

Make resolveTS and friends take a const TSCtx& instead of many separate args. Make the version of TypeStructure::resolve that's called from class.cpp take just the necessary bits of the Class::Const as separate args, and have all versions of TypeStructure::resolve package their args up in a TSCtx instead of a fake/copied Class::Const.

Reviewed By: oulgen

Differential Revision: D18705254

fbshipit-source-id: 531a242b9794a694752b0783caccc1b662963fea