watchman telemetry
Summary:
The goal of this diff is to gather telemetry about watchman invocations.

Previous diffs already gathered all the information we need. This diff merely logs it.

Note: there are already a bunch of watchman telemetry events inside EventLogger, invoked by the "watchman" module in src/watchman/watchman. That's a very different thing. Its job is to maintain the persistent watchman subscription and listen to notifications. It's something that we intend to eliminate. So, I want to deprecate those things, in favor of new ones.

Reviewed By: arxanas

Differential Revision: D18756525

fbshipit-source-id: 3bdb7d523cfb17eb660cb7bcad8f5037dcddb416