Add equality functions for typedef_type and fun_elt.
Summary: Comparing decls by printing them to strings feels elegant, but it's quite lossy, and makes it easy to think that two decls are identical when they're actually different.  We already have nice comparison methods for consts and classes, but we'll need to add new ones for typedefs and fun_elts.

Reviewed By: 2BitSalute

Differential Revision: D18605916

fbshipit-source-id: 8df1d83605e1a60911b1c96ea19758b16fb01624