Parallelize GDB indexer
Summary:
The goal of this diff is to further speed up the GDB index generation. This is done by using a pool of workers and introducing parallel versions of CU visit functions.

With this change, indexing HHVM in DBG mode
* Thread count 2: 34.4 seconds
* Thread count 4: 21 seconds
* Thread count 8: 14.6 seconds
* Thread count 16: 11 seconds
* Thread count 32: 10.1 seconds
* Thread count 48: 10.3 seconds

With this change, indexing HHVM in DBGO mode
* Thread count 2: 24 seconds
* Thread count 4: 14.5 seconds
* Thread count 8: 9.2 seconds
* Thread count 16: 7.1 seconds
* Thread count 32: 6.6 seconds
* Thread count 48: 6.4 seconds

With this change, indexing HHVM in opt-hhvm-lto mode
* Thread count 2: 16.7 seconds
* Thread count 4: 10.5 seconds
* Thread count 8: 7 seconds
* Thread count 16: 5.5 seconds
* Thread count 32: 5.1 seconds
* Thread count 48: 5.1 seconds

iwantsomemoreparallelism

Reviewed By: markw65

Differential Revision: D18774861

fbshipit-source-id: 1d76def6f711ff38a0173d5109880a00e03ef9b1