Implement VregSet more efficiently
Summary:
The current implementation of VregSet is a sorted vector of Vregs,
which works surprisingly well. However we can do even better. Instead
represent the VregSet using a more traditional bitset
approach. Represent the set using an array of blocks. Each block
contains 256 bits. The index of the bit corresponds to the Vreg it
represents. All operations on the VregSet are done on a block
basis. This allows for the compiler to unroll the loops and use SIMD
operations where appropriate. As an additional optimization, if the
VregSet would use only a single block, store it "inline" without
needing any heap allocation. This means that VregSet can contain all
Vregs up to 256 without needing heap allocations (this includes all
physical registers).

Reviewed By: mofarrell

Differential Revision: D18799903

fbshipit-source-id: 0483b369914962de43287aa24c8dce89781bb0b5