Use better DecRefProfile for inline returns
Summary:
When we dec-ref locals for an inlined region, we currently use the last (usually, the coldest) translation's profiles, because these are the last ones emitted. Instead, we should look at all predecessor SrcKeys and take the hottest one.

We could do one better by merging these profiles, but at least this is an improvement over the current state.

Reviewed By: arnabde03

Differential Revision: D18432067

fbshipit-source-id: b25f471f35adb6adc863594e34953aed1d098ca4