bump up LSP test timeout
Summary:
The sIDE integration tests cause spurious failure messages 50%-80% of the time.

My experience with ad-hoc use of sIDE is that it often stalls for no apparent reason before it's able to answer the first request. I've seen it often stall for 10+ seconds.

The current timeout is only 5 seconds. I think that's certainly the cause of the problem. If I'm experience slowdowns even on my uncontested devserver, then the slowdowns on overloaded Sandcastle machines will certainly be worse.

I spent ages eliminating flakiness from the Flow LSP integration tests. To do that for Flow, I bumped up the timeouts to far more than seemed reasonable. I settled on ~2 minutes.

This means that it takes us 2 minutes before we get signal on "my test failed to get any answer from the sIDE" when we're developing locally. This seems a fair tradeoff to make given (1) we're not doing much sIDE work right now, (2) when we are doing sIDE, most failure's aren't of the form that it simply fails to deliver an answer; instead they're of the form that it crashes or delivers a wrong answer, (3) reliability of tests is absolute paramount.

Separately, we really have to hunt down and characterize the cause of sIDE slow initial response. That's a separate topic I'm working on.

So, for the meantime, this diff just bumps up the timeout considerably.

Reviewed By: arxanas

Differential Revision: D18827048

fbshipit-source-id: b193f1dac15c47881fc5e7808e4132c86f474b3b