Use a special data-structure instead of std::deque for cached operands
Summary:
In vasm-graph-color, originally we used std::vector to store all the
CachedOperands. However this was problematic because when the vector
resized, it would invalidate all references to the contents, which
caused subtle bugs. So, we switched to using std::deque instead, which
doesn't have the invalidation issue. However, this caused a
performance regression. Access in a std::deque is slower than a
std::vector, and we access CachedOperands a lot.

Introduce a new data-structure called NonInvalidatingVector. It has a
very basic API, similar to std::vector. Like std::vector, access is
O(1), but it never invalidates references to its contents. The
trade-off for this is certain restrictions on element types, and
possible increased memory usage. These are fine for vasm-graph-color,
however.

Use NonInvalidatingVector to store CachedOperands.

Reviewed By: mofarrell

Differential Revision: D18820600

fbshipit-source-id: fd12405203bf5d0c95825c314c3d35603649c82b