Speed up penalties calculation in vasm-graph-color
Summary:
The calculation of penalty vectors involves walking the unit, and
calculating which Vregs are alive at the same time as any physical
registers (as well as which spill slots are simultaneously live). This
was done using standard liveness analysis, walking backwards and
maintaining a set of live Vregs.

I noticed this logic was taking longer then expected. Once I
investigated, I realized it was because we were always unioning in the
defs set of the current instruction into the live set, then processing
that. Once we are done, we then remove the defs out of the live
set. The problem is, many times the def set contains a lot of physical
registers (because of call instructions and such). Since we skip over
physical registers in the processing loop, this means we spend a lot
of effort unioning in a lot of physical registers, iterating over
them, and then skipping them (all inside an inner loop). This is all
wasted work.

Do it a bit differently to be more efficient. Enforce that the live
set will never contain physical registers (this is cheap to do with
the new VregSet implementation). The live set is the set of Vregs
which will be assigned a penalty, which means we don't care about
physical registers in it. Also, don't union in the def set to the live
set. We're going to remove it immediately after the processing
loop. Instead we just add the across set (we need to do this because
they interfer with the defs). The only complication this causes is for
the spill slot interference calculation. In that case we need to loop
over both the live set and the def set.

Reviewed By: mofarrell

Differential Revision: D18827834

fbshipit-source-id: 73dff260bddde76c381574d208ceaf406ee64091